Added JExcelApi
Test to use as a reader for any of the jobs pulled from the Postgres
Server. Can hopefully use to pull any type of selection of data. That's
why I'm using Excel instead of just simple CSV as it has more
functionality and will hopefully make things easier in the long-run.
http://www.vogella.com/tutorials/JavaExcel/article.html